Split SHA256Functor out
