+++ lib/Basics/SHA256Functor.h
@@ -1,0 +1,39 @@
+////////////////////////////////////////////////////////////////////////////////
+/// DISCLAIMER
+///
+/// Copyright 2014-2022 ArangoDB GmbH, Cologne, Germany
+/// Copyright 2004-2014 triAGENS GmbH, Cologne, Germany
+///
+/// Licensed under the Apache License, Version 2.0 (the "License");
+/// you may not use this file except in compliance with the License.
+/// You may obtain a copy of the License at
+///
+///     http://www.apache.org/licenses/LICENSE-2.0
+///
+/// Unless required by applicable law or agreed to in writing, software
+/// distributed under the License is distributed on an "AS IS" BASIS,
+/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and
+/// limitations under the License.
+///
+/// Copyright holder is ArangoDB GmbH, Cologne, Germany
+///
+/// @author Dr. Frank Celler
+////////////////////////////////////////////////////////////////////////////////
+
+#include <string>
+
+////////////////////////////////////////////////////////////////////////////////
+/// @brief functor for generating a SHA256.  Use with TRI_ProcessFiles.
+///        you need to wrap your TRI_SHA256Functor object within std::ref().
+////////////////////////////////////////////////////////////////////////////////
+struct TRI_SHA256Functor {
+  TRI_SHA256Functor();
+  ~TRI_SHA256Functor();
+
+  bool operator()(char const* data, size_t size) noexcept;
+
+  std::string finalize();
+
+  void* _context;
+};
